# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Upwork AI Automation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List all proposals
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, approved, sent, rejected]
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: List of proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get a single proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "404":
          description: Not found
    patch:
      operationId: updateProposal
      tags: [proposals]
      summary: Update a proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    delete:
      operationId: deleteProposal
      tags: [proposals]
      summary: Delete a proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /proposals/{id}/approve:
    post:
      operationId: approveProposal
      tags: [proposals]
      summary: Approve a proposal for sending
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
...
components:
  schemas:
...
    Proposal:
      type: object
      required: [id, jobId, content, status, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        jobTitle:
          type: ["string", "null"]
        content:
          type: string
        coverLetter:
          type: ["string", "null"]
        bidAmount:
          type: ["number", "null"]
        estimatedDuration:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, approved, sent, rejected]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ProposalUpdate:
      type: object
      properties:
        content:
          type: string
        coverLetter:
          type: string
        bidAmount:
          type: number
        estimatedDuration:
          type: string
        status:
          type: string
          enum: [draft, approved, sent, rejected]
...